requirements.txt
# Core dependencies
streamlit>=1.28.0
boto3>=1.34.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# AWS Strands Agent SDK
strands-agents>=1.0.0
strands-agents-tools>=0.2.8

# Development dependencies
pytest>=7.4.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0